requirements.txt
psycopg2
python-dotenv
celery[redis]
